fix: update architecture
