+++ src/index.html
@@ -11,7 +11,7 @@
 </head>
 
 <body>
-    <script type="module" src="scripts/index.js"></script>
+    <script type="module" src="scripts/app.js"></script>
 </body>
 
 </html>
