+++ src/scripts/index.js
@@ -1,21 +1,8 @@
-import { createElement } from './app.js';
-import { Keyboard } from './keyboard.js';
-
-const createPage = () => {
-  const body = document.querySelector('body');
-  const title = 'Virtual Keyboard';
-  const description = 'Клавиатура создана в операционной системе Windows';
-  const language = 'Для переключения языка комбинация: левыe ctrl + alt';
-
-  const wrapper = createElement('div', '', 'wrapper');
-  const keyboard = new Keyboard();
-  wrapper.append(createElement('h1', title, 'title'));
-  wrapper.append(createElement('textarea', '', 'textarea'));
-  wrapper.append(keyboard.createKeyboard());
-  wrapper.append(createElement('p', description, 'description'));
-  wrapper.append(createElement('p', language, 'language'));
-
-  body.append(wrapper);
+const createElement = (el, innerHTML, ...classes) => {
+  const element = document.createElement(el);
+  element.classList.add(...classes);
+  element.innerHTML = innerHTML;
+  return element;
 };
 
-createPage();
+export { createElement };
